@@ -1,16 +1,14 @@
 package com.ansen.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.TextView;
 
-import com.ansen.recyclerview.R;
+import androidx.recyclerview.widget.RecyclerView;
 
+import com.ansen.recyclerview.R;
 import java.util.List;
 
 /**
