@@ -1,18 +1,16 @@
 package com.ansen.recyclerview;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SimpleItemAnimator;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.AdapterView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.ansen.adapter.RecyclerHeadFootViewAdapter;
-import com.ansen.adapter.RecyclerViewAdapter;
 import com.ansen.util.DividerItemDecoration;
 
 import java.util.ArrayList;
@@ -35,7 +33,7 @@ public class HeadFootActivity extends AppCompatActivity {
 
         initData();
 
-        recyclerView= (RecyclerView) findViewById(R.id.recyclerview);
+        recyclerView= findViewById(R.id.recyclerview);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));//设置布局管理器
         recyclerView.addItemDecoration(new DividerItemDecoration(this));//添加分割线
         recyclerView.setAdapter(adapter=new RecyclerHeadFootViewAdapter(this,datas));
